changed service implementation to consume grapql github api (v4)
