+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.levelup.bibangamba.githubusers">
+    package="com.levelup.bibangamba.githubusersGraphQL">
 
     <uses-permission android:name="android.permission.INTERNET" />
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
@@ -11,17 +11,17 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name="com.levelup.bibangamba.githubusers.view.MainActivity">
+        <activity android:name="com.levelup.bibangamba.githubusersGraphQL.view.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name="com.levelup.bibangamba.githubusers.view.DetailActivity">
+        <activity android:name="com.levelup.bibangamba.githubusersGraphQL.view.DetailActivity">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="com.levelup.bibangamba.githubusers.view.MainActivity" />
+                android:value="com.levelup.bibangamba.githubusersGraphQL.view.MainActivity" />
         </activity>
     </application>
