+++ app/src/main/graphql/com.levelup.bibangamba.githubusersGraphQL/getKenyanJavaDevsFromGithub.graphql
@@ -1,0 +1,25 @@
+query get40KenyanJavaDevsQuery{
+    search(type:USER, query:"location:Nairobi language:JAVA", first: 40){
+        userCount
+        edges{
+            node{
+                ... on User{
+                    login
+                    bio
+                    avatarUrl
+                    company
+                    repositories{
+                        totalCount
+                    }
+                    url
+                    followers{
+                        totalCount
+                    }
+                    following{
+                        totalCount
+                    }
+                }
+            }
+        }
+    }
+}
