--- app/src/main/java/com/levelup/bibangamba/githubusers/util/CheckNetworkConnection.java
@@ -1,19 +1,0 @@
-package com.levelup.bibangamba.githubusers.util;
-
-import android.content.Context;
-import android.net.ConnectivityManager;
-import android.net.NetworkInfo;
-
-public class CheckNetworkConnection {
-    private Context context;
-    public CheckNetworkConnection(Context context){
-        this.context = context;
-    }
-
-    public boolean isNetworkAvailable(){
-        ConnectivityManager connectivityManager
-                = (ConnectivityManager) context.getSystemService(Context.CONNECTIVITY_SERVICE);
-        NetworkInfo activeNetworkInfo = connectivityManager.getActiveNetworkInfo();
-        return activeNetworkInfo != null && activeNetworkInfo.isConnected();
-    }
-}
