--- app/src/main/java/com/levelup/bibangamba/githubusers/view/GithubUserDetailsView.java
@@ -1,7 +1,0 @@
-package com.levelup.bibangamba.githubusers.view;
-
-import com.levelup.bibangamba.githubusers.model.GithubUsers;
-
-public interface GithubUserDetailsView {
-    void githubUserInformationFetchComplete(GithubUsers githubUser);
-}
