--- app/src/main/java/com/levelup/bibangamba/githubusers/view/GithubUsersView.java
@@ -1,9 +1,0 @@
-package com.levelup.bibangamba.githubusers.view;
-
-import java.util.List;
-
-import com.levelup.bibangamba.githubusers.model.GithubUsers;
-
-public interface GithubUsersView {
-    void githubUsersHaveBeenFetchedAndAreReadyForUse(List<GithubUsers> githubUsers);
-}
