+++ app/src/main/java/com/levelup/bibangamba/githubusersGraphQL/view/GithubUserDetailsView.java
@@ -1,0 +1,7 @@
+package com.levelup.bibangamba.githubusersGraphQL.view;
+
+import com.levelup.bibangamba.githubusersGraphQL.model.GithubUsers;
+
+public interface GithubUserDetailsView {
+    void githubUserInformationFetchComplete(GithubUsers githubUser);
+}
