+++ app/src/main/java/com/levelup/bibangamba/githubusersGraphQL/view/GithubUsersView.java
@@ -1,0 +1,10 @@
+package com.levelup.bibangamba.githubusersGraphQL.view;
+
+import java.util.List;
+
+import com.levelup.bibangamba.githubusersGraphQL.Get40KenyanJavaDevsQuery;
+import com.levelup.bibangamba.githubusersGraphQL.model.GithubUsers;
+
+public interface GithubUsersView {
+    void githubUserInformationFetchFromGraphQLComplete(List<Get40KenyanJavaDevsQuery.Edge> edges);
+}
