+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:id="@+id/main_activity_constraint_layout"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.levelup.bibangamba.githubusers.view.MainActivity">
+    tools:context="com.levelup.bibangamba.githubusersGraphQL.view.MainActivity">
 
     <ProgressBar
         android:id="@+id/github_users_progress_bar"
