+++ build.gradle
@@ -1,14 +1,14 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 
 buildscript {
-    
+    ext.apolloClientVersion = "1.0.0-alpha5"
     repositories {
         google()
         jcenter()
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.2.1'
-        
+        classpath 'com.android.tools.build:gradle:3.3.2'
+        classpath "com.apollographql.apollo:apollo-gradle-plugin:$apolloClientVersion"
 
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
@@ -53,4 +53,6 @@ ext {
     powerMockVersion="1.6.2"
     mockitoVersion="1.10.19"
     uiautomatorVersion = "2.2.0"
+    apolloClientVersion = "1.0.0-alpha5"
+    logginInterceptor = "3.13.1"
 }
